+++ app/src/main/java/com/willowtree/vocable/customviews/CategoryButton.kt
@@ -1,10 +1,10 @@
-package com.example.eyespeak.customviews
+package com.willowtree.vocable.customviews
 
 import android.content.Context
 import android.speech.tts.TextToSpeech
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatRadioButton
-import com.example.eyespeak.utils.VocableTextToSpeech
+import com.willowtree.vocable.utils.VocableTextToSpeech
 import kotlinx.coroutines.*
 
 /**
