+++ app/src/main/java/com/willowtree/vocable/customviews/PointerView.kt
@@ -1,4 +1,4 @@
-package com.example.eyespeak.customviews
+package com.willowtree.vocable.customviews
 
 import android.content.Context
 import android.util.AttributeSet
