+++ app/src/main/java/com/willowtree/vocable/customviews/SettingsButton.kt
@@ -1,9 +1,9 @@
-package com.example.eyespeak.customviews
+package com.willowtree.vocable.customviews
 
 import android.content.Context
 import android.content.Intent
 import android.util.AttributeSet
-import com.example.eyespeak.settings.SettingsActivity
+import com.willowtree.vocable.settings.SettingsActivity
 
 /**
  * A subclass of VocableButton that will open SettingsActivity when interacted with
