+++ app/src/main/java/com/willowtree/vocable/facetracking/FaceTrackingViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.eyespeak.facetracking
+package com.willowtree.vocable.facetracking
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
