+++ app/src/test/java/com/willowtree/vocable/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.eyespeak
+package com.willowtree.vocable
 
 import org.junit.Test
